+++ scripts/database/otrs-schema.maxdb.sql
@@ -693,7 +693,7 @@ CREATE INDEX index_time_accountin62 ON time_accounting (ticket_id)
 // ----------------------------------------------------------
 CREATE TABLE ticket_watcher (
     ticket_id BIGINT NOT NULL,
-    user_id BIGINT NOT NULL,
+    user_id INTEGER NOT NULL,
     create_time timestamp NOT NULL,
     create_by INTEGER NOT NULL,
     change_time timestamp NOT NULL,
